Major change on JSClosure
- Add `JSOneshotClosure` which provide an ability to invoke a closure
  at most once.
- Remove `JSClosure.init(_ body: @escaping ([JSValue]) -> Void)`
  overload, which forces users to write result type in closure.
- Add JSClosure lifetime test suites
